refactor: move token connection to serverSide
